Moved util to validate
